+++ docker-run.sh
@@ -13,4 +13,4 @@ docker run -d --name="Zoneminder" \
 -e MULTI_PORT_END="0" \
 -v "/mnt/cache/appdata/Zoneminder":"/config":rw \
 -v "/mnt/cache/appdata/Zoneminder/data":"/var/cache/zoneminder":rw \
-zoneminder
+dlandon/zoneminder.machine.learning